refactor(spf): split text-track loader and host provider
Phase 3 of decoupling text-track orchestration from DOM. The prior
single behavior in dom/behaviors/load-text-track-cues.ts did two jobs:
orchestrate state transitions (DOM-free), and manage actor lifecycle
tied to an HTMLMediaElement (DOM-bound). Splitting these concerns
lets the orchestration live host-agnostic in media/behaviors/.

Changes:

- media/types: add MediaElementWithTextTracks — MediaElementLike plus
  an iterable textTracks of {id: string}. HTMLMediaElement satisfies
  it structurally.

- media/behaviors/load-text-track-cues.ts (new): host-agnostic
  orchestrator. Reads actors from owners; never creates them. State
  machine drops 'setting-up' — actor presence is now just another
  precondition. DOM-free enforced by the media subproject tsconfig.

- dom/behaviors/provide-text-track-actors.ts (new): DOM-side provider
  that creates/destroys TextTracksActor and the segment-loader actor
  on mediaElement mount/unmount, writing them to owners. Replaces
  the actor lifecycle logic previously buried in the orchestrator.

- dom/behaviors/load-text-track-cues.ts: deleted. Owner management
  moves to the provider; orchestration moves to media.

- hls-engine: imports loadTextTrackCues from media/behaviors, and
  adds provideTextTrackActors to the composition list alongside it.

- dom/index.ts: replaces the loadTextTrackCues export with
  provideTextTrackActors (the loader is now re-exported by media).

- Tests updated: setup composes both provider and loader, types the
  shared owners signal as the intersection so it satisfies each
  behavior's shape (HTMLMediaElement narrows MediaElementWithTextTracks).

Benefits:
- The orchestration can now be reused by non-DOM hosts (worker,
  test fake, future MoQ engine) that provide their own actor factories.
- Actor lifecycle is explicit as its own behavior — easier to reason
  about and to substitute in alternate engines.
- media/behaviors/ now has a real occupant beyond the pure-function
  behaviors; validates the directory earning its name.

All green: build (163 files), tests (707), typecheck, workspace
check, lint.
